+++ tests/examples/simple_cond_conf/ignored_rules/rules_simple_cond_ignored_rules.yaml
@@ -13,7 +13,7 @@ rules:
         action_parameters:
           value: true
       IGNORED_RULE_2:
-        simple_condition: input.age>1
+        simple_condition: input.dummy>1
         action: set_admission
         action_parameters:
           value: true
@@ -22,3 +22,5 @@ rules:
         action: set_admission
         action_parameters:
           value: false
+
+parsing_error_strategy: raise
